+++ Runtime/ObjPool/IPool.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+
+namespace OcUtility
+{
+    // public interface IPool<T> where T : MonoBehaviour, IPoolMember<T>
+    // {
+    //     int InitialCount { get; }
+    //     Transform Folder { get; }
+    //     T Call(in Vector3 position, in Quaternion rotation);
+    //     T Call(in Vector3 position);
+    //     void Return(T member);
+    // }
+}