+++ Runtime/ObjPool/OcPool.cs
@@ -6,42 +6,47 @@ using Object = UnityEngine.Object;
 
 namespace OcUtility
 {
-    public class OcPool : IDisposable
+    public class OcPool<T> : IDisposable where T : MonoBehaviour, IPoolMember<T>
     {
-        static Dictionary<PoolMember, OcPool> GlobalPool;
+        static Dictionary<IPoolMember<T>, OcPool<T>> GlobalPool;
         
         static OcPool()
         {
 #if UNITY_EDITOR
-            Application.quitting += () => GlobalPool = new Dictionary<PoolMember, OcPool>();      
+            Application.quitting += () => GlobalPool = new Dictionary<IPoolMember<T>, OcPool<T>>();      
 #endif
-            GlobalPool = new Dictionary<PoolMember, OcPool>();
+            GlobalPool = new Dictionary<IPoolMember<T>, OcPool<T>>();
         }
 
-        public static OcPool MakePool(PoolMember source, int count)
+        public static OcPool<T> MakePool(T source, int count, Transform folder)
         {
-            OcPool targetPool;
+            OcPool<T> targetPool;
             if (GlobalPool.ContainsKey(source))
             {
                 targetPool = GlobalPool[source];
                 targetPool.AddMember(count);
             }
-            else targetPool = new OcPool(source, count);
+            else targetPool = new OcPool<T>(source, count, folder);
 
             return targetPool;
         }
 
+        public static OcPool<T> MakePool(T source, int count)
+        {
+            return MakePool(source, count, null);
+        }
+
         /// <summary> 글로벌 풀에서 해당 소스 기반의 풀에서 Call을 반환함. 해당 소스의 풀이 없을 경우 null을 반환함. </summary>
-        public static PoolMember Call(PoolMember source, in Vector3 position, in Quaternion rotation)
+        public static T Call(T source, in Vector3 position, in Quaternion rotation)
         {
             if (GlobalPool.ContainsKey(source))
             {
                 var targetPool = GlobalPool[source];
                 
                 // 씬이 변경되는 등의 이유로 참조가 Missing 상태일 경우 다시 풀을 재생성함.
-                if (targetPool._folder == null)
+                if (targetPool.Folder == null)
                 {
-                    GlobalPool[source] = MakePool(source, targetPool._initialCount);
+                    GlobalPool[source] = MakePool(source, targetPool.InitialCount);
                     targetPool = GlobalPool[source];
                 }
                 return targetPool.Call(in position, in rotation);
@@ -50,77 +55,142 @@ namespace OcUtility
             return null;
         }
 
-        public OcPool FindPool(PoolMember source)
+        public static T Call(T source, in Vector3 position)
+        {
+            return Call(source, in position, Quaternion.identity);
+        }
+
+        public static OcPool<T> FindPool(T source)
         {
             if (GlobalPool.ContainsKey(source)) return GlobalPool[source];
 
             return null;
         }
 
-        public OcPool(PoolMember source, int count)
+        public static void DisposeAll()
         {
-            _folder = new GameObject().transform;
+            PoolDisposer.DisposeAll();
+            GlobalPool = new Dictionary<IPoolMember<T>, OcPool<T>>();
+        }
+
+        public static void Print()
+        {
+            Debug.Log("||========================================||");
+            Debug.Log($"||Global Pool<{typeof(T).Name}> | Count : {GlobalPool.Count}||");
+            foreach (var ocPool in GlobalPool)
+            {
+                Debug.Log($"||PoolType : {ocPool.Key.GetType()} | Count : {ocPool.Value.TotalCount}||");    
+            }
+            Debug.Log("||=========================================||");
+        }
+        
+        
+        // Non-Static. =======
+        
+        public OcPool(T source, int count, Transform folder)
+        {
+            _isOverridenFolder = folder != null;
+            Folder = folder == null ? new GameObject().transform : folder;
+
             _source = source;
             GlobalPool[source] = this;
-            _sleepingMembers = new Queue<PoolMember>();
-            _allMembers = new List<PoolMember>();
-            _initialCount = count;
-            AddMember(count, true);
+            _sleepingMembers = new Queue<T>();
+            _activeMembers = new List<T>();
+            InitialCount = count;
+            AddMember(count);
+            
+            PoolDisposer.RegisterPool(this);
         }
         
+        public Transform Folder { get; }
+        public int TotalCount => _sleepingMembers.Count + _activeMembers.Count;
+        public int InitialCount { get; }
+
+        bool _isOverridenFolder;
+        T _source;
+        Queue<T> _sleepingMembers;
+        List<T> _activeMembers;
         
-        // Non-Static.
-        readonly Transform _folder;
-        PoolMember _source;
-        Queue<PoolMember> _sleepingMembers;
-        List<PoolMember> _allMembers;
-        readonly int _initialCount;
-        bool HasEnoughMember(int count)
+        public void AddMember(int count)
         {
-            return _allMembers.Count * 0.1f > count && _sleepingMembers.Count * 0.5f  > count;
-        }
-        /// <summary> 오브젝트 풀에 멤버를 추가함. 기본적으로 전체 풀의 10%미만, 대기 풀의 50% 미만이면 추가하지 않고 넘어감. </summary>
-        public void AddMember(int count, bool forceToAdd = false)
-        {
-            if(!forceToAdd && HasEnoughMember(count)) return;
             for (int i = 0; i < count; i++)
             {
-                var gao = Object.Instantiate(_source, _folder, true);
+                var gao = Object.Instantiate(_source, Folder, true);
                 gao.Pool = this;
                 _sleepingMembers.Enqueue(gao);
-                _allMembers.Add(gao);
                 gao.gameObject.SetActive(false);
             }
 
-            _folder.name = $"{_source.name}_Pool [{_allMembers.Count:###,###}]";
+            if(!_isOverridenFolder) Folder.name = $"{_source.name}_Pool [{TotalCount:###,###}]";
         }
 
-        public PoolMember Call(in Vector3 position, in Quaternion rotation)
+        public T Call(in Vector3 position, in Quaternion rotation)
         {
-            if(_sleepingMembers.Count == 0) AddMember(_initialCount);
+            if(_sleepingMembers.Count == 0) AddMember(InitialCount);
 
             var member = _sleepingMembers.Dequeue();
             member.transform.SetPositionAndRotation(position, rotation);
             member.WakeUp();
+            _activeMembers.Add(member);
             return member;
         }
 
-        public void Return(PoolMember member)
+        public T Call(in Vector3 position)
+        {
+            return Call(in position, Quaternion.identity);
+        }
+
+        public T FindActiveMember(Predicate<T> predicate)
+        {
+            for (int i = 0; i < _activeMembers.Count; i++)
+            {
+                var m = _activeMembers[i];
+                if (predicate.Invoke(m)) return m;
+            }
+
+            return null;
+        }
+
+        public T FindSleepingMember(Predicate<T> predicate)
+        {
+            var toArray = _sleepingMembers.ToArray();
+            for (int i = 0; i < _activeMembers.Count; i++)
+            {
+                var m = toArray[i];
+                if (predicate.Invoke(m)) return m;
+            }
+
+            return null;
+        }
+
+        public void Foreach(Action<T> action)
+        {
+            for (int i = 0; i < _activeMembers.Count; i++)
+            {
+                var m = _activeMembers[i];
+                action.Invoke(m);
+            }
+
+            var toArray = _sleepingMembers.ToArray();
+            for (int i = 0; i < _sleepingMembers.Count; i++)
+            {
+                var m = toArray[i];
+                action.Invoke(m);
+            }
+        }
+
+        public void Return(T member)
         {
             _sleepingMembers.Enqueue(member);
+            _activeMembers.Remove(member);
         }
 
         public void Dispose()
         {
             _source = null;
             _sleepingMembers = null;
-            var count = _allMembers.Count;
-            for (int i = 0; i < count; i++)
-            {
-                Object.Destroy(_allMembers[0]);
-            }
-
-            _allMembers = null;
+            _activeMembers = null;
+            if(!_isOverridenFolder) Object.Destroy(Folder.gameObject);
         }
     }
 }